Create chapter_1_SNP_script
This script used the full E. vaginatum transcriptome VCF file to transform Phytochrome transcripts sequence data in the form of a FASTA file. Parameters were incorporated to only focus on mutations that were likely to be real, rather than sequencing errors within the transcriptome. An output directory is created with transcript names and sequences of transcripts that fall among the given parameters.