Create chapter_2_translate
This script was developed to be used with the program SciRoKo 3.4, this program output file has information on microsatellite markers that occur in a given FASTA file. The script reads in the FASTA file and translates these genes on each reading frame, then storing where the longest coding region occurs. The positions of the microsatellites are compared with the coding region positions, and an output file is created with only the microsatellites that occur inside the coding regions. The script can also be easily altered the look at microsatellites outside the coding regions.

Script Arguments:
-f FASTA file input name
-s SciRoKo output file
-l SSR (microsatellite) minimum repeat (ex. dinucleotide = 2)
-out_file Output file
-out_fasta FASTA file output