Update chapter_1_SNP
This script was developed to parse genes that meet specific parameters for mutation quality from a transcriptome using a VCF and FASTA file. The script reads in the sequences from the FASTA file and mutation quality and frequency from the VCF file, if the mutations fall within the given parameters, the sequence data is transformed with the mutation and written to an output directory. This output directory contains only the gene names and sequences that qualify with the threshold parameters.

Script Arguments:
-f FASTA file input
-v VCF file input
-ad Allele depth (for mutation quality)
-sd Sample depth (for mutation frequency)
-o Output directory name